fix worker deletion
